requirements.txt
# UI
streamlit>=1.28.0

# OCR
pytesseract>=0.3.10
easyocr>=1.7.1
pdf2image>=1.16.3
Pillow>=10.0.0

# Google Gemini API
google-generativeai>=0.3.0

# データベース
SQLAlchemy>=2.0.0

# ユーティリティ
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# テスト
pytest>=7.4.0 
